django_p/requirements.txt
django>=5.1,<5.2
djangorestframework
pyyaml
requests
django-cors-headers
